Move active channel logic to slice
- Moved the logic for setting the active channel from ChatPage to channelsSlice.
Previosly, ChatPage had useEffect, which set the default channel as active if the user was in the active channel and deleted it. This logic is now implemented in extraReducers of the channels slice.
So, removed useEffect in ChatPage.
- In channelsSlice:
Added addMatcher for addChannel - the new chanel automatically becomes active.
Added addMatcher for removeChannel - if the active channel is deleted, the default channel becomes active.